+++ examples/alchemy-daapp/src/surfaces/onboarding/OnboardingController.ts
@@ -1,13 +1,14 @@
+import { withAlchemyGasManager } from "@alchemy/aa-alchemy";
+import {
+  SimpleSmartContractAccount,
+  SmartAccountProvider,
+  createPublicErc4337Client,
+  type SimpleSmartAccountOwner
+} from "@alchemy/aa-core";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { encodeFunctionData } from "viem";
-import {
-  createPublicErc4337Client,
-  SimpleSmartContractAccount,
-  type SimpleSmartAccountOwner,
-  SmartAccountProvider,
-  alchemyPaymasterAndDataMiddleware,
-} from "@alchemy/aa-core";
 import { useAccount, useNetwork } from "wagmi";
+import { localSmartContractStore } from "~/clients/localStorage";
 import { NFTContractABI } from "../../clients/nftContract";
 import {
   DAAppConfiguration,
@@ -21,7 +22,6 @@ import {
   initialStep,
   metaForStepIdentifier,
 } from "./OnboardingDataModels";
-import { localSmartContractStore } from "~/clients/localStorage";
 
 async function pollForLambdaForComplete(
   lambda: () => Promise<boolean>,
@@ -97,10 +97,14 @@ const onboardingStepHandlers: Record<
       throw new Error("No entrypoint address was found");
     }
     const entryPointAddress = context.entrypointAddress;
-    const baseSigner = new SmartAccountProvider(
+    let baseSigner = new SmartAccountProvider(
       appConfig.rpcUrl,
       context.entrypointAddress!,
-      context.chain!
+      context.chain!,
+      undefined,
+      {
+        txMaxRetries: 60,
+      }
     ).connect((provider: any) => {
       if (!context.owner) {
         throw new Error("No owner for account was found");
@@ -113,15 +117,16 @@ const onboardingStepHandlers: Record<
         rpcClient: provider,
       });
     });
+
+    
     const smartAccountAddress = await baseSigner.getAddress();
     if (context.useGasManager) {
-      const smartAccountSigner = await baseSigner.withPaymasterMiddleware(
-        alchemyPaymasterAndDataMiddleware({
-          provider: baseSigner.rpcClient,
-          policyId: appConfig.gasManagerPolicyId,
-          entryPoint: entryPointAddress,
-        })
-      );
+      const smartAccountSigner = withAlchemyGasManager(baseSigner, {
+        provider: baseSigner.rpcClient,
+        policyId: appConfig.gasManagerPolicyId,
+        entryPoint: entryPointAddress,
+      });
+  
       return {
         nextStep: OnboardingStepIdentifier.MINT_NFT,
         addedContext: {